Move Eth types out of utils
