+++ orchestrator/ethereum_gravity/src/types.rs
@@ -1,0 +1,5 @@
+use ethers::prelude::*;
+use std::sync::Arc;
+
+pub type EthSignerMiddleware = SignerMiddleware<Provider<Http>, LocalWallet>;
+pub type EthClient = Arc<EthSignerMiddleware>;
