+++ orchestrator/orchestrator/src/main_loop.rs
@@ -20,7 +20,8 @@ use deep_space::client::ChainStatus;
 use deep_space::error::CosmosGrpcError;
 use deep_space::private_key::PrivateKey as CosmosPrivateKey;
 use deep_space::{Contact, Msg};
-use ethereum_gravity::utils::{EthClient, get_gravity_id};
+use ethereum_gravity::types::EthClient;
+use ethereum_gravity::utils::get_gravity_id;
 use ethers::{prelude::*, types::Address as EthAddress};
 use gravity_proto::gravity::query_client::QueryClient as GravityQueryClient;
 use gravity_utils::ethereum::bytes_to_hex_str;
