+++ orchestrator/orchestrator/src/oracle_resync.rs
@@ -1,5 +1,5 @@
 use deep_space::address::Address as CosmosAddress;
-use ethereum_gravity::utils::EthClient;
+use ethereum_gravity::types::EthClient;
 use ethers::prelude::*;
 use ethers::types::Address as EthAddress;
 use gravity_proto::gravity::query_client::QueryClient as GravityQueryClient;
